add config class to composite service
