+++ microservices/product-composite-service/src/main/java/com/yavlash/microservices/composite/product/config/ApplicationConfiguration.java
@@ -1,0 +1,82 @@
+package com.yavlash.microservices.composite.product.config;
+
+import com.yavlash.microservices.composite.product.services.ProductCompositeIntegration;
+import io.swagger.v3.oas.models.ExternalDocumentation;
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Contact;
+import io.swagger.v3.oas.models.info.Info;
+import io.swagger.v3.oas.models.info.License;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.actuate.health.CompositeReactiveHealthContributor;
+import org.springframework.boot.actuate.health.ReactiveHealthContributor;
+import org.springframework.boot.actuate.health.ReactiveHealthIndicator;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+import java.util.LinkedHashMap;
+import java.util.Map;
+
+@Configuration
+@ComponentScan("com.yavlash")
+public class ApplicationConfiguration {
+    @Value("${api.common.version}")
+    String apiVersion;
+    @Value("${api.common.title}")
+    String apiTitle;
+    @Value("${api.common.description}")
+    String apiDescription;
+    @Value("${api.common.termsOfService}")
+    String apiTermsOfService;
+    @Value("${api.common.license}")
+    String apiLicense;
+    @Value("${api.common.licenseUrl}")
+    String apiLicenseUrl;
+    @Value("${api.common.externalDocDesc}")
+    String apiExternalDocDesc;
+    @Value("${api.common.externalDocUrl}")
+    String apiExternalDocUrl;
+    @Value("${api.common.contact.name}")
+    String apiContactName;
+    @Value("${api.common.contact.url}")
+    String apiContactUrl;
+    @Value("${api.common.contact.email}")
+    String apiContactEmail;
+
+    /**
+     * Will exposed on $HOST:$PORT/swagger-ui.html
+     *
+     * @return the common OpenAPI documentation
+     */
+    @Bean
+    public OpenAPI getOpenApiDocumentation() {
+        return new OpenAPI()
+                .info(new Info().title(apiTitle)
+                        .description(apiDescription)
+                        .version(apiVersion)
+                        .contact(new Contact()
+                                .name(apiContactName)
+                                .url(apiContactUrl)
+                                .email(apiContactEmail))
+                        .termsOfService(apiTermsOfService)
+                        .license(new License()
+                                .name(apiLicense)
+                                .url(apiLicenseUrl)))
+                .externalDocs(new ExternalDocumentation()
+                        .description(apiExternalDocDesc)
+                        .url(apiExternalDocUrl));
+    }
+
+    @Autowired
+    ProductCompositeIntegration integration;
+
+    @Bean
+    ReactiveHealthContributor coreServices() {
+        final Map<String, ReactiveHealthIndicator> registry = new LinkedHashMap<>();
+        registry.put("product", () -> integration.getProductHealth());
+        registry.put("recommendation", () -> integration.getRecommendationHealth());
+        registry.put("review", () -> integration.getReviewHealth());
+        return CompositeReactiveHealthContributor.fromMap(registry);
+    }
+}
